Moved loggers into separate file.
