+++ Metrics.js
@@ -1,4 +1,4 @@
-import { sw_log, sw_error_log } from './index';
+import { sw_log, sw_error_log } from './loggers';
 import { get, set } from './db';
 
 export default class Metrics {
@@ -34,16 +34,6 @@ export default class Metrics {
           lastRun,
         }).catch(sw_error_log);
       }
-      await set('metrics', hash, {
-        ...metrics,
-        timesElapsed: metrics.timesElapsed.concat(timeElapsed),
-      }).catch(sw_error_log);
     }
-    sw_log(`Time elapsed: ${timeElapsed}`);
-    sw_log(`isCached: ${this.isCached}`);
   }
 }
-
-//     if (!notCached) notCached = (Date.now() - uncachedStart)/1000;
-//     else cachedEnd = (Date.now() - cachedStart)/1000;
-//     console.log("Not cached Speed:", notCached, "Cached speed:", cachedEnd)
