+++ index.js
@@ -1,12 +1,4 @@
-export const sw_log = (...strings) =>
-  strings.forEach((msg) => {
-    console.log(`%c ${msg}`, 'color: green; font-weight: bold');
-  });
-
-export const sw_error_log = (...strings) =>
-  strings.forEach((msg) => {
-    console.log(`%c ${msg}`, 'color: red; font-weight: bold');
-  });
+import { sw_log, sw_error_log } from './loggers';
 
 export const register = () => {
   if (navigator.serviceWorker) {
